change style of start page and fix pagination prob
